calendar_sync.py
import re
from datetime import datetime, timedelta
from googleapiclient.discovery import build
from google_auth_oauthlib.flow import InstalledAppFlow
from google.auth.transport.requests import Request

# Google Calendar API setup (replace placeholders with your credentials)
SCOPES = ['https://www.googleapis.com/auth/calendar']
creds = None
# ... (load or refresh token, see Google API documentation)
service = build('calendar', 'v3', credentials=creds)

def parse_schedule(filename):
    with open(filename, 'r') as file:
        content = file.read()

    # Regex patterns for different schedule elements
    month_pattern = r"(\w+)\."  # e.g., "June."
    week_pattern = r"(\d{2})\."  # e.g., "23." (week number)
    date_pattern = r"([A-Za-z]{3}-\d{1,2})\."  # e.g., "Mon-17."
    time_pattern = r"(\d{2}:\d{2})"  # e.g., "09:00"
    task_pattern = r"([-\s]\s+)(.+)"  # e.g., "- Reply to emails"

    events = []
    tasks = []

    current_date = None
    for line in content.splitlines():
        # Extract month
        month_match = re.match(month_pattern, line)
        if month_match:
            month_name = month_match.group(1)
            continue

        # Extract week number (optional)
        week_match = re.match(week_pattern, line)

        # Extract date
        date_match = re.match(date_pattern, line)
        if date_match:
            day_name, day_num = date_match.group(1).split('-')
            # ... (Calculate current_date using month_name, day_name, and day_num) 

        # Extract time and task
        time_match = re.match(time_pattern, line)
        if time_match:
            time_str = time_match.group(1)
            task_match = re.match(task_pattern, line[len(time_str):])
            if task_match:
                task = task_match.group(2)
                # ... (Create event or task with current_date, time_str, and task)

    return events, tasks

def sync_to_google(events, tasks):
    for event in events:
        # ... (Create or update event in Google Calendar using the API)

    for task in tasks:
        # ... (Create or update task in Google Tasks using the API)

if __name__ == '__main__':
    filename = 'test-schedule.txt'  # Replace with your actual file name
    events, tasks = parse_schedule(filename)
    sync_to_google(events, tasks)
